+++ gmailer/urls.py
@@ -4,7 +4,6 @@ from gmailer import views
 
 app_name = 'gmailer'
 urlpatterns = [
-    path('', views.paths, name='paths'),
     path('auth', views.auth, name='auth'),
     path('verify', views.verify, name='verify'),
     path('revoke', views.revoke, name='revoke'),