+++ gmailer/views.py
@@ -6,7 +6,7 @@ from django.urls import reverse
 from gmailer.gmail import mailer
 
 def auth(request):
-    url, request.session['oauth_state'] = mailer.authorize(request)
+    url, request.session['oauth_state'] = mailer.authorize()
     return redirect(url)
 
 def verify(request):
@@ -15,13 +15,8 @@ def verify(request):
 
 def revoke(request):
     mailer.revoke()
-    return redirect('gmailer:auth')
-    return JsonResponse({
-        'message': 'Successfully revoked the API service',
-        'urls': mailer.urls })
+    return JsonResponse({'message': 'Successfully revoked the API service'})
 
 def test_send_mail(request):
     mailer.test_mail()
-    return JsonResponse({
-        'message': 'Successfully send test mail',
-        'urls': mailer.urls })
+    return JsonResponse({'message': 'Successfully send test mail'})