+++ docs/conf.py
@@ -13,10 +13,17 @@
 import re
 import os
 import sys
+import django
 
-sys.path.insert(0, os.path.abspath('..'))
-
+sys.path.insert(0, os.path.abspath('../'))
 from gmailer import __version__
+
+# Specify settings module
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'test_project.settings')
+
+# Setup Django
+django.setup()
+
 
 # -- Project information -----------------------------------------------------
 
@@ -59,4 +66,4 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = []